requirements.txt
streamlit
crewai
crewai_tools
pymupdf
python-docx
IPython
